==> Code/5-team-clustering.ipynb <==
%matplotlib inline
import warnings
warnings.filterwarnings('ignore')
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
from IPython.display import display, HTML
from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = "all"

pd.options.display.max_columns = None

gl = pd.read_csv("./data/final_gl.csv")

season_end_performance = gl[(gl.year>=2019)&(gl.year<=2023)].groupby(['HomeTeam','year'])[['season_end_rank','season_end_w_l_ratio']].max().reset_index()

# Wide format results
season_end_performance = season_end_performance.pivot_table(index='HomeTeam',columns='year')

# flatten columns
flat_cols = [str(col[1])+"_"+col[0] for col in season_end_performance.columns]
season_end_performance.columns = flat_cols
season_end_performance = season_end_performance.reset_index()


attendance_stats = gl[(gl.year>=2018)&(gl.year<=2022)].groupby('HomeTeam')['Attendance'].agg(['mean','std']).reset_index()
attendance_stats.rename({'mean':'3yr_attedance_mean', 'std':'3yr_attedance_std'},inplace=True)

cluster_df = season_end_performance.merge(attendance_stats, on='HomeTeam')

from sklearn.cluster import KMeans

# Initialize the KMeans algorithm with 5 clusters
kmeans = KMeans(n_clusters=4)

cluster_df.fillna(0,inplace=True)
# Fit the KMeans model to the data
kmeans.fit(cluster_df[cluster_df.columns[1:]])

# Get the cluster labels for each point
labels = kmeans.labels_

# Get the centroids of each cluster
centroids = kmeans.cluster_centers_

# print cluster characteristics
pd.concat([pd.DataFrame([0,1,2,3]), pd.DataFrame(centroids)], axis=1)
# 3, 1, 2, 0

cluster_df.columns

cluster_df['label'] = labels

cluster_df.sort_values('label')
